fix(frontend): no phantom draft when opening a CLI-pushed script
